+++ Changelog.md
@@ -5,10 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased [1.3.52] 
 
-## Unreleased [1.3.51] - 
+### Improved
 
-### Added
+- Paddings for table now are smaller on lower screen sizes
+
+## [1.3.51] - 2024-10-03
 
 ### Fixed